refactor: split block runtime blocks helpers
